+++ dmapiclient/search.py
@@ -116,15 +116,9 @@ class SearchAPIClient(BaseAPIClient):
                 raise
         return None
 
-    def search_services(self, index, q=None, page=None, id_only=False, **filters):
+    def search(self, index, doc_type, q=None, page=None, id_only=False, **filters):
         response = self._get(
-            self.get_search_url(index=index, doc_type='services', q=q, page=page, id_only=id_only, **filters)
-        )
-        return response
-
-    def search_briefs(self, index, q=None, page=None, id_only=False, **filters):
-        response = self._get(
-            self.get_search_url(index=index, doc_type='briefs', q=q, page=page, id_only=id_only, **filters)
+            self.get_search_url(index=index, doc_type=doc_type, q=q, page=page, id_only=id_only, **filters)
         )
         return response