+++ CHANGELOG.md
@@ -9,7 +9,7 @@ PR: [#116](https://github.com/alphagov/digitalmarketplace-apiclient/pull/116)
 `search_services` method on the search-api-client has been generalized to `search` and now takes a `doc_type` argument.
 This allows more than just services to be searched.
 
-`aggregate_services` method on the search-api-client has been generalized to `aggregate_docs` to allow for aggregating
+`aggregate` method on the search-api-client has been generalized to `aggregate` to allow for aggregating
 briefs too. It takes an extra argument, `doc_type`.
 
 Old:
@@ -20,7 +20,7 @@ search_api_client.search_services(index, q=q, page=page, id_only=id_only, **filt
 
 New:
 ```
-search_api_client.aggregate_docs(index=index, doc_type='services', q=q, aggregations=aggregations)
+search_api_client.aggregate(index=index, doc_type='services', q=q, aggregations=aggregations)
 search_api_client.search(index, doc_type, q=q, page=page, id_only=id_only, **filters)
 ```